Example moved to one file
Refs: #6718
