--- packages/samples/react/src/scenarios/horizontal-scrollbar-advanced/DemoPage.tsx
@@ -1,17 +1,0 @@
-import * as React from 'react';
-import { SampleDescription } from '../../components/SampleDescription';
-import MainLayout from './MainLayout';
-import DemoTable from './Table';
-
-function DemoPage() {
-	return (
-		<>
-			<SampleDescription></SampleDescription>
-			<MainLayout>
-				<DemoTable />
-			</MainLayout>
-		</>
-	);
-}
-
-export default DemoPage;
