--- packages/samples/react/src/scenarios/horizontal-scrollbar-advanced/MainLayout.tsx
@@ -1,20 +1,0 @@
-import * as React from 'react';
-import Navigation from './Navigation';
-import TabsLayout from './TabsLayout';
-
-import './layout.css';
-
-type MainLayoutProps = {
-	children: React.ReactNode;
-};
-
-function MainLayout({ children }: MainLayoutProps) {
-	return (
-		<div className="mainlayout">
-			<Navigation className="nav-area" />
-			<TabsLayout className="content" demoContent={children} />
-		</div>
-	);
-}
-
-export default MainLayout;
