--- packages/samples/react/src/scenarios/horizontal-scrollbar-advanced/Table/columns.ts
@@ -1,7 +1,0 @@
-export const columnDefinitions: unknown = [
-	{ key: 'name', label: 'Name', textAlign: 'left', width: '400px' },
-	{ key: 'species', label: 'Species', textAlign: 'left', width: '400px' },
-	{ key: 'habitat', label: 'Habitat', textAlign: 'left', width: '400px' },
-	{ key: 'diet', label: 'Diet', textAlign: 'left', width: '400px' },
-	{ key: 'lifespan', label: 'lifespan', textAlign: 'right', width: '400px' },
-];
