--- packages/samples/react/src/scenarios/horizontal-scrollbar-advanced/Table/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Table';
